pyproject.toml
[build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
requires = ["setuptools>=61.0.0", "wheel"]

[project]
name = "dolfin_pyvista_adapter"
version = "0.1.0"
description = "Adapter for Pyvista for arbitrary order (discontinuous) Lagrange with Legacy Dolfin."
authors = [{name = "Jørgen S. Dokken", email = "dokken@simula.no"}]
license = {file = "LICENSE"}
readme = "README.md"
dependencies = ["fenics-dolfin", "pyvista"]


[project.optional-dependencies]
test = [
    "pytest",
]
dev = [
    "pdbpp",
    "ipython",
"mypy",
    "flake8",
]
docs = [
   "jupyter-book",
   "jupytext",
   "panel"
]
all = [
   "dolfin_pyvista_adapter[test]",
   "dolfin_pyvista_adapter[dev]",
   "dolfin_pyvista_adapter[docs]"
]

[tool.pytest.ini_options]
addopts = [
   "--import-mode=importlib",
]
testpaths = [
    "tests"
]

[tool.mypy]
ignore_missing_imports = true
# Folders to exclude
exclude = [
    "docs/",
    "build/"
]
# Folder to check with mypy
files = [
    "src",
    "tests"
]
